IIntervalCollection.change disallow undefined endpoint (#17094)
## Description

IIntervalCollection.change allows one of the endpoints to be undefined.
This means that it should stay wherever it is. But this can
unintentionally result in the end being before start (especially in
conjunction with remote edits). For compat, ops are still allowed to
leave an endpoint unspecified.

There doesn't seem to be any usages of this functionality anyway.

## Breaking Changes

Usages of the change method that passed in undefined endpoint should now
specify the current endpoint position.

## Reviewer Guidance

I updated the skipped fuzz tests and verified that they all fail with
assert 0x54e or RedBlackTree.nodeRemove "Cannot read properties of
undefined (reading 'left')". Let me know if there's anything else I
should verify to ensure it's the known bugs.