tree2:  Cleanup use of InMemoryStoredSchemaRepository (#17132)
## Description

InMemoryStoredSchemaRepository is a specific implementation of
SchemaRepository, and a lot of our code needlessly required that
specific implementation due to how it was typed. This made implementing
the forestWithContent helper messy, so was fixed.

While doing this, places that cloned InMemoryStoredSchemaRepository were
adjusted accordingly, and commented with regards to how why shoul
dhandeling editing. The SchemaRepository type only exists to provide
editability and change events ontop of SchemaData, so places that clone
it should be clear about what they do with edits or not clone it or use
SchemaData instead.

Additionally the policy in InMemoryStoredSchemaRepository is no longer
needed (since default schema handling was changed a while ago) so it was
removed as a simplification.