tree2: Decouple Forest and Schema. (#17139)
## Description

This reduces coupling between forest and schema. This simplifies the
codebase and is particularly useful as schema is being reworked and this
avoids having to address if forest is exporting the stored or view
schema.

This decoupling is not complete since a schema is still provided to the
forest when cloning and when building a chunked forest. This allows the
forest to use the schema for heuristic purposes (like inferring good
chunk shapes): this will still need some clarification for stored vs
view, but becomes simpler after this change.

## Breaking Changes

Forest no longer exports the schema, nor invalidates when schema
changes.