build(client-utils): Make private=false and use correct test names (#17143)
I haven't received objections to using the fluid-internal scope for this
package, so I'm removing its private property so there aren't any issues
when we try to release it.

The test names for the hash tests were the old "common-utils" so I fixed
those. I also removed an unused previous version dependency.