Remove bindToContext API (#16911)
## Description
This change removes the `bindToContext` API that has been deprecated for
a long time. Some of the partners were dependent on it but that
depdendency has been removed, so its safe to remove it now.

## Breaking
This is a breaking change. The `bindToContext` API has been deprecated
since a long time.


[AB#4434](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4434)