Simplfied SummarizerNode:refreshLatestSummary logic (#17106)
## Description
#16657 removed the logic to refresh a summary's state by downloading a
snapshot. With it gone, summarizer node (and rest of the system) only
updates state from a summary that it was tracking. This PR simplifies
the logic in `SummarizerNode::refreshLatestSummary` to account for this.
It also renames the properties in the returned object to give the caller
more context on what happened and what action it can take.

Additionally, it removes logging the `PendingSummaryNotFound` telemetry.
It is currently logged every time the summarizer loaded for the summary
that is loaded from. The reason is that it processes the ack for the
summary is loaded from and since the summary was not generated by this
instance of the summarizer, this event is logged. Instead, added a
property `pendingSummaryTracked` to the refreshLatestSummary_end event.


[AB#4435](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4435)