Wait for pending ops to be saved before summarization (#16465)
## Description
There can be pending (unacked) ops in the summarizer client during
summarization. This leads to eventually inconsistent and incorrect
summaries. #16961 added failing summarization if there are pending ops
**_after_** summary has been generated.
This change adds failing summarization if there are pending ops
**_before_** generating summary. Additionally, it will fail for some
time (default 1 second) for pending ops to be processed before starting
to generate summary. The idea behind this change is that if there are
pending ops before we even start to generate summary, it will fail
anyway. So, might as well fail early and the wait is an optimization to
give it a chance to start with a clean state.

## Key points
- The logic to wait for pending ops is behind a feature flag so that we
can experiment with it and look at data before turning it on.
- Waiting for pending ops to be saved will happen with a timeout so that
we don't end up waiting for a long time. The behavior when the timeout
expires is to continue with summarization to keep the behavior same as
today. Otherwise, it can lead to documents being stuck, i.e., not
summarizing (maybe for a long time). We can tweak this behavior based on
the data we see with the experiment.


[AB#4903](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4903)