Switch the whole repo to use c8 instead of nyc (#17181)
The client release groups already switch to c8 in #5990. This PR will
switch the rest of the release group and includes all the per package
"test:coverage" script, which are only used for local development (since
in the CI we call c8 that covers pnpm command that activate the "test:*"
script instead of calling "test:coverage".