Deprecate old logger types (#17149)
We added "Ext" logging types to be used when instrumenting, to add
support for logging flat arrays and objects directly without having to
JSON.stringify (the logger does it for you). This functionality was in
new copies of old types, with "Ext" in the name.

This commit deprecates the old types, and includes renaming some other types
to align on the "Base" nomenclature for types used in public API surface
(which don't include the extended types mentioned above).