docs: Fix `@example` comments to adhere to TSDoc spec (#17207)
The TSDoc spec notes that an `@example` comment with contents on the
same line as the tag should treat the first line as the title of that
example. We had a number of comments that had example body contents on
the same line as the tag. This PR adjusts contents to better adhere to
the spec.

It also makes some line spacing improvements to enhance readability, and
makes some other misc. comment improvements.

`@example` spec: https://tsdoc.org/pages/tags/example/