Connect in 'read' mode by default (#17210)
## Description

When loading a container, it needs to connect in 'read' mode. The only
exception is when the container is loaded with a pending state
containing stashed ops. This change also provides an escape hatch in
case someone really needs 'write mode', via a feature gate. The feature
gate is also currently used to work around some existing test failures.