[port 6.1] Server deps update 1.0.0->1.0.1 (#17244)
## Description

Ports the update to server dependencies from 1.0.0 to 1.0.1
(https://github.com/microsoft/FluidFramework/pull/17085) to the release
branch for v2int/6.1. Done because we ported it to 6.0
[here](https://github.com/microsoft/FluidFramework/pull/17242) and need
the subsequent minors to have this change too. 6.3 has it naturally.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).