feat(api-markdown-documenter): Correctly handle `@example` comment titles (#17217)
The TSDoc spec notes that an `@example` comment with contents on the
same line as the tag should treat the first line as the title of that
example. This PR updates the documenter library to respect this in its
output.

`@example` comments with titles now output contents like the following
(example in Markdown, heading level is arbitrary):

```markdown
# Example: <title>

<example-comment-body>
```

### Example

For example, a comment like the following:

```typescript
/**
 * ...
 * @example Error case
 * 
 * An example error includes...
 */
```

Would produce documentation output like the following (example in
Markdown, heading level is arbitrary):

```markdown
# Example: Error case

An example error includes...
```

`@example` spec: https://tsdoc.org/pages/tags/example/

### Breaking Change

This PR is API breaking in that it adds the `isLiteral` and `isParent`
properties to the `DocumentationNode` interface. This is used to power
generic tree walking, like that required to implement the (very hacky)
title-stripping logic.

### Other changes

* Updates dependencies on:
  * `@microsoft/api-documenter`
  * `@microsoft/api-extractor-model`
* Updates test assets to illustrate this `@example` case
* Bumps package version from `0.7.2` to `0.8.0` in preparation for
publishing