refactor(api-markdown-documenter): Rename the `createChildContentSections` configuration option to `createDefaultLayout` (#17291)
This config option evolved somewhat organically, and the name has become
fairly obsolete / misleading. Ultimately, it is a layout option that
applies as the default for all API items. As an implementation detail,
it takes in and renders item-specific elements (previously referred to
as "child contents") within the layout, but this does not accurately
reflect the user-facing semantics of the option.

### Breaking Change

This PR changes the public API by renaming the configuration option, and
is breaking.