[Summary-stress] Added basic summarization stress test workflow (#17272)
## Description
The goal of the stress test is to be able to stress the summarization
system in ways like real world scenarios. We want to be able to run our
system through real user-like activities over time and validate that
things work as expected.
This is version 1 of the summarizer stress tests and it attempts to do
the following:
- Have activities at regular intervals so as to trigger summarizations.
- Grow the number objects (data stores and DDSes) in the container over
time so that summaries are bigger and take more time.
- Change certain parts of the container to trigger incremental
summaries. For instance, only change some data stores. Within a data
store, only change some or no DDSes.

## Structure and details
The hierarchical structure of objects is as follows:

![image](https://github.com/microsoft/FluidFramework/assets/45832642/b9a62989-d2c8-4d49-95a6-8256896c44e8)

This is what the objects do:
- At a regular interval, the root data store does one of the following:
  - Does nothing
  - Creates a new child data store.
- At a regular interval, the child data stores do one of the following:
  - Does nothing
  - Creates a new DDS
  - Changes one DDS
  - Changes two DDSes