remove(api-markdown-documenter): AlertNode (#17292)
Removes the `AlertNode` class, which never should have been a part of
this API. It was added as a short term solution to meet the needs of the
fluidframework.com website build. This library's API surface is now
sufficiently configurable that this detail can be moved into the website
build, rather than being a part of this API, which is intended to be
general purpose.

The existing usages (`@deprecated` and `@beta` notices) have been
updated to use simpler output syntax. Additionally, support for `@alpha`
was added, simply because it was convenient to do so while refactoring.

Additionally, `alpha` and `beta` notices in tables were moved from the
"Summary" cell to the "Alerts" cell (comma-separated with deprecation
alerts).

* The test assets illustrate the differences in output.

### Breaking Change

This removes a class from the public API, and is therefore breaking for
consumers.