[GC stress] Refactored the stress test infra to support running different work loads (#17302)
## Reviewer guidance
This is in a test branch and is experimental. The goal is to move this
to the main branch eventually.

## Description
Refactored the stress test infra to support running different workloads.
Currently, it has 3 workloads - default, gc and summary.
Here are the details:
- Created a `workloads` directory that contains a subdirectory for each
workload.
- The workload directory must contain the following two files:
- fluidExport.ts - This should export a data store factory with the name
`fluidExport`. The factory must implement
`IProvideFluidDataStoreFactory`.
The data store that the factory creates must implement `ITestRunner`
interface. This is the entry point into the test work load that the test
runner calls.
- testConfig.json - This contains the configurations that the test runs
with. The test will choose one of the profiles based on the `--profile`
param.
- The test has a default work load under the `default` workload
directory. New workloads can be added and configured to be run via npm
commands.
- Updated the yaml pipeline to run GC and summary tests for ODSP and
tinylicious.

## Follow ups
- Update the yaml pipelines to run the default workload as well (similar
to the one in main).
- Move the GC pass / fail logic into the GC test runner.
- Make the pass / fail logic configurable.