[GC-stress] Add ability for test runners to abort and report failure (#17332)
Added support in the stress test infra for test runners to abort and
report the failure. The failures would show up in the `Test` tab in ADO
pipelines.
Here is how it works:
- `ITestRunner:run` function returns a `TestRunResult` object with the
following type. If the test wants to report failure, it should set
`abort: true` and provied an `errorCode` that is a non-zero number. This
will result in this test runner reported as failed.
``` typescript
/**
 * The result of running the test on a specific runner.
 */
export type TestRunResult =
  | {
        /** Abort and fail the runner */
	abort: true;
	/** The errorCode that the runner exits with. Should be non-zero. */
	errorCode: number;
    }
  | {
	/** The runner should not be aborted */
	abort: false;
	/** Whether the runner is done. If this is false, the runner reruns the test */
	done: boolean;
    };
```
- An example usage of this is `RootDataObject` in `gcDataStores.ts`
which returns a failure when certain events are logged.