Make entryPoint required (#17040)
## Breaking Changes

### provideEntryPoint is required

The optional `provideEntryPoint` method has become required on a number
of constructors. A value will need to be provided to the following
classes:

-   `BaseContainerRuntimeFactory`
-   `RuntimeFactory`
-   `ContainerRuntime` (constructor and `loadRuntime`)
-   `FluidDataStoreRuntime`

See
[testContainerRuntimeFactoryWithDefaultDataStore.ts](https://github.com/microsoft/FluidFramework/tree/main/packages/test/test-utils/src/testContainerRuntimeFactoryWithDefaultDataStore.ts)
for an example implemtation of `provideEntryPoint` for ContainerRuntime.
See
[pureDataObjectFactory.ts](https://github.com/microsoft/FluidFramework/tree/main/packages/framework/aqueduct/src/data-object-factories/pureDataObjectFactory.ts#L83)
for an example implementation of `provideEntryPoint` for
DataStoreRuntime.

Subsequently, various `entryPoint` and `getEntryPoint()` endpoints have
become required. Please see
[containerRuntime.ts](https://github.com/microsoft/FluidFramework/tree/main/packages/runtime/container-runtime/src/containerRuntime.ts)
for example implementations of these APIs.

For more details, see
[Removing-IFluidRouter.md](https://github.com/microsoft/FluidFramework/blob/main/packages/common/core-interfaces/Removing-IFluidRouter.md)

### ContainerRuntimeFactory constructors have changed

The following class constructors have been changed to allow for better
flexible in arguments passed:

-   `BaseContainerRuntimeFactory`
-   `ContainerRuntimeFactoryWithDefaultDataStore`
-   `RuntimeFactory`

They now use a single object for constructor params. Example change to
be made:

```ts
// Old
new BaseContainerRuntimeFactory(
	myRegistryEntries,
	myDependencyContainer,
	myRequestHandlers,
	myRuntimeOptions,
	myProvideEntryPoint,
);

// New
new BaseContainerRuntimeFactory({
	registryEntries: myRegistryEntries,
	dependencyContainer: myDependencyContainer,
	requestHandlers: myRequestHandlers,
	runtimeOptions: myRuntimeOptions,
	provideEntryPoint: myProvideEntryPoint,
});
```

### initializeEntryPoint changed to provideEntryPoint

The naming of `initializeEntryPoint` has been changed to
`provideEntryPoint`. Please change the property name in relevant calls
to `ContainerRuntime.loadRuntime(...)`.


[AB#4992](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4992)

## Reviewer Guidance

Most of this PR is providing value for `provideEntryPoint` in tests.
I'll point out that this is because `provideEntryPoint` is now required
on the ContainerRuntimeFactoryWithDefaultDataStore class.
Alternatively, we could leave it as optional on that class and provide
this getDefaultFluidObject method as the default.

Also looking for feedback on if the `getEntryPoint()` API should keep
its return type as `Promise<FluidObject | undefined>` or if it's fine to
remove the `undefined` portion.