Only stash blobs when closing (#17230)
This change makes getPendingLocalState to stash blobs only in
closeAndGetPendingLocalState case (notifyImminentClosure = true).
getPendingLocalState without closing can't stash blobs given that their
reference could not be included in the pending state, having as a result
a blob handle without any reference, dds that uses it.

Some tests relied on being able to read pending attachment blobs which
is not possible now, so I changed some asserts and ways to confirm
correct behavior. There is one test we're losing here "blob is acked
after upload". Given that we can't collect pending blobs if they are not
attached, there is no way to know if the stashed blob is purely acked. I
substituted that test with "removes pending blob when waiting for blob
to be attached" which indirectly checks the same thing, given that only
blobs who are attacked and acked are removed.