Bump default GC version in code to 3 (#17371)
The old version was 2, with an override for 3 available, and utilized by
some internal early-adopters of GC to regenerate the GC Data, after some
Summarizer fixes were made that were found to be corrupting the GC data.

Now 3 should be the default value, and if/when additional bugs are found
that require regenerating the GC data for existing containers, we can
bump the version to 4.