fix(api-markdown-documenter): Restore `getApiItemTransformationConfigurationWithDefaults` in public API (#17413)
This function was removed previously, as it is intended as an
implementation detail. Unfortunately, since the public facing
API-Item-related utility functions require a complete configuration to
work (rather than a partial, user-provided config), this API is needed
for users to call any of those APIs.

Added a TODO to the README about updating the APIs to accept a partial
configuration, but that will be a larger project.