TS5: Fix 'type-test-generator' to unblock bumping 'build-tools' (#17416)
The current prelease version of 'type-test-generator' contains a change
to TypeOnly<T> that strips nominal typing from strings and numbers.
While seemingly more permissive, this introduced 2 type test failures in
'@fluidframework/file-driver'.

I believe the error message points to the _prototype_ of this ctor fn:
```ts
export const FileSnapshotWriterClassFactory = <TBase extends ReaderConstructor>(Base: TBase) =>
  class extends Base implements ISnapshotWriterStorage {
    ...
  }
```
And the symptoms indicate that a Boolean type is causing the TypeOnly
comparison to fail, although the compiler does not indicate the specific
field (it just reports that the 'prototype' field is not assignable.)

Further modifying TypeOnly<T> to preserve the original typing of boolean
fixes this issue.  Other primitives (bigint and symbol) are preserved as well
for good measure.  The 'null' and 'undefined' values cannot be branded
(intersection returns never) and do not require special treatment.

---------

Co-authored-by: Craig Macomber (Microsoft) <42876482+CraigMacomber@users.noreply.github.com>