Deprecate request pattern (#17359)
## Upcoming Changes

### IContainer's and IDataStore's IFluidRouter capabilities are
deprecated

`IFluidRouter` and `request({ url: "/" })` on `IContainer` and
`IDataStore` are deprecated and will be removed in a future major
release. Please migrate all usage to the appropriate `getEntryPoint()`
or `entryPoint` APIs.

See
[Removing-IFluidRouter.md](https://github.com/microsoft/FluidFramework/blob/main/packages/common/core-interfaces/Removing-IFluidRouter.md)
for more details.

### IFluidRouter and IProvideFluidRouter deprecated

`IFluidRouter` and `IProvideFluidRouter` have been deprecated. Please
remove all usages of these interfaces and migrate to the new
`entryPoint` pattern.

### Various request related APIs have been deprecated

Please remove all calls to the following functions and instead use the
new `entryPoint` pattern:

-   `requestFluidObject`
-   `requestResolvedObjectFromContainer`
-   `getDefaultObjectFromContainer`
-   `getObjectWithIdFromContainer`
-   `getObjectFromContainer`


[AB#4986](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4986)