feat(merge-tree sequence): change the direction local references look to slide to (#17286)
Change the direction local references look to slide to. This change is
breaking in terms of behavior for existing intervals, but should better
match the desired and intuitive behavior of reference sliding. This
change breaks quite a number of tests, and so is disabled for tests
except for those involving stickiness, as this change directly ties into
stickiness.

The exact change is that: rather than going `forward ?? backward ??
detached` (or `backward ?? forward ?? detached`), we slide `forward ??
detached` _or_ `backward ?? detached`. The cases of detached can also be
substituted for the special endpoint segments that exist immediately
before and after the bounds of the tree/string.


[ADO#5205](https://dev.azure.com/fluidframework/internal/_workitems/edit/5205/)