Inject the telemetry token into the Devtools browser extension during build (#17269)
## Description
In order for the Devtools browser extension to successfully send usage
telemetry to Aria/Kusto, it needs a token. We don't want to put that
token in source control if we can avoid it, to minimize the risk of
someone obtaining it and abusing it. We should find a way to inject this
value into the extension code during build (both local and CI).

[AB#5167](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5167)

This change will also upload the injected key and built Devtools browser
extension into build client pipeline artifact.

## Example
![Screenshot 2023-09-12
110307](https://github.com/microsoft/FluidFramework/assets/114451900/5b8f8ea3-960f-4f42-a0de-73c2ab9f58d1)

---------

Co-authored-by: Kajari Ghosh <ghoshkaj@gmail.com>
Co-authored-by: Joshua Smithrud <54606601+Josmithr@users.noreply.github.com>