tree2: Remove FieldKindTypes (#17457)
## Description

`FieldKindTypes` is a union of the currently default supported
implementations of `FieldKind`.
This change replaces use of it with `FieldKind`, which makes the code a
bit more conventional, but also more generic in that it will will work
with future FieldKinds as well as eventual legacy kinds for back compat
more simply.

Also cleans up a few typos in the domains files.


## Breaking Changes

`FieldKindTypes` has been removed. Usages of it should be replaced with
`FieldKind` and handle cases where the kind is an unexpected type.