build(docs): Add optional title for @examples section in docs (#17482)
The api-markdownd-documenter package exposes a utility function called
`createExamplesSection` (in Helpers.ts), which generates the section
contents describing all `@example` comments on an API item. Currently,
the heading text "Examples" is hard-coded for that section. For FF.com,
we wish to use "Usage", instead of "Examples".

To accomplish this, we will need to update the createExamplesSection
function to accept an optional parameter for the heading text.

Later (after a new version has been published and consumed in the
website build), we can specify "Usage" as the value for this function
call.