build(docs): Add optional title for @throws section in docs (#17481)
The api-markdownd-documenter package exposes a utility function called
`createThrowsSection` (in Helpers.ts), which generates the section
contents describing all `@throws` comments on an API item. Currently,
the heading text "Throws" is hard-coded for that section. For FF.com, we
wish to use "Error Handling", instead of "Throws".

To accomplish this, we will need to update the createThrowsSection
function to accept an optional parameter for the heading text.

Later (after a new version has been published and consumed in the
website build), we can specify "Error Handling" as the value for this
function call.