ci: publish packages to feeds based on settings (#17477)
This PR unifies the logic that decides what feeds a given package can be
published to and uses that logic in the CI pipeline to determine what
packages to publish to a feed.

It does this by creating a file with a list of packages. One file is
created for each of the feed kinds - official, internal, and
internal-test. These files are then used in the bash script that
publishes the packages. Only packages listed in the files are published,
so this effectively filters out packages that shouldn't be published to
a particular feed, including examples and most fluid-internal packages.

Since the policy-check logic is used to determine whether a package
should be published, as we update that policy, the publishing pipeline
should benefit from those changes and "just work."

CI run from a test branch testing this change:
https://dev.azure.com/fluidframework/internal/_build/results?buildId=193778&view=results