Summarization: Wrap any errors thrown from mixed in summary handler as DataProcessingError (#17442)
When a consumer of FF uses `mixinSummaryHandler` to run additional code
before summarization, they introduce risk to the Summarization system
where that code could throw and cause summarization to fail. These
failures are completely outside the control of FF, but can be
catastrophic for a container since persistent failure to summarize can
lead to effective document corruption (e.g. ODSP blocks all changes once
there are 10k unsummarized ops).

So we want to classify these errors as DataProcessingErrors with the
codepath `mixinSummaryHandler` mentioned so we and our partners can have
immediate insight when these errors arise. Currently, a premier 1P
partner is facing a large number of these errors and we are trying to
clarify the telemetry signal to be more actionable.