Add unboxing iterators to EditableTree2 (#17451)
## Description

This replaces the current iterator behavior of the editable tree types
to be specific to the type of tree object, and in general favors
unboxing. The existing iteration behavior (which keeps nodes and fields
boxed) has been moved to a method called `[boxedIterator]()`.

## Breaking Changes

The behavior of node and field iterators has been moved to
`[boxedIterator]()`.