refactor(azure): Enable no-abbreviations eslint rule and fix violations (#17562)
Abbreviations present an accessibility concern for new developers on the
team, and developers whose native language is not American English. This
PR enables the
[unicorn/prevent-abbreviations](https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/prevent-abbreviations.md)
rule for the Azure packages, and fixes violations where it would not
introduce API-breaking changes.

This PR can be considered a prototype, as I would personally like to see
this rule enabled globally across the repo to make our code more
accessible both for developers on the team and our customers.