Update logs for KafkaRunner and producer's error handlers (#17592)
## Description

1. The KafkaRunner metric logs often contain exceptions that might be
bubbled up from the doc corruption flow or the restart flow. Adding a
property in those logs to identify such exceptions so that we can filter
them and understand the impact of remaining exceptions. Similarly
updated the logs for Kafka producer's error handler.

2. Adding some logs in partitionManager to debug an issue where restart
might be triggered from more than one flow.

## Reviewer Guidance

Using `errorLabel` as the new property name to indicate which flow the
error is coming from. Open to suggestions for a better name.

---------

Co-authored-by: Shubhangi Agarwal <shuagarwal@microsoft.com>