build: Upgrade pipelines to Node18 + upgrade nvm recommendation (#17389)
Update the recommended versions of Node.js for developers to use, as
well as the version used by CI, from version 14 to 18.

This change made some scenarios in some e2e tests hang, so a fix for
skipping them had to be included.
Details:
0.58.x doesn't have the fix to routerlicious-driver in this PR:
https://github.com/microsoft/FluidFramework/pull/8913
(also see related issue for more context on the problem:
https://github.com/microsoft/FluidFramework/issues/9163)
This causes this test to hang while loading container2, as the snapshot
is over 16KB.