Fix pipeline script error when there are test errors (#17187)
When there are script error, the "Log Test error" steps would have a
script error `[: too many arguments` because of lack of quotes.

While fixing this, I improved how it works:
- Simplify the shell script to just invoking the `report-parser.js` and
having the JS script output the ADO messages.
- Change to use ADO logissue message to get it into the UI.

Now if there are test errors, it will show up in the run UI instead of
having to dig into the Test tab. For example:


![image](https://github.com/microsoft/FluidFramework/assets/12101885/32eea400-a6a6-4c49-89ac-95302c31fc62)

Also add the pipeline template as a dependency so that change to the
template will trigger the pipelines that uses it.