ci: Centralize feed publishing and unify style (#17608)
Broken out of #17578.

The CI pipeline logic for publishing to a feed was split. Most feeds
were published to in individual jobs with a reusable template, but then
there was a case where the npmrc was used to publish \@fluid-internal
packages to a separate feed. I assume this was originally to save an
extra npm publish step (let npm do the work of sorting based on the
npmrc) but it was confusing and difficult to debug.

This PR unifies all the CI logic regarding publishing to feeds to use
the same style. That is, we generate a list of packages per feed, then
use those lists in the pipeline to decide what to publish where. The
npmrc gets the same values for every scope - the lists are the source of
truth.

I also renamed some of the workflow parameters to better reflect what
they are. E.g. `feedUrl` instead of `feedName`.

I think we could also remove the splitting of packages by
scoped/unscoped. That step seems unnecessary now that we have an
explicit list for each feed. That can be done at a later time so isn't
included here.