Don't stringify shared tree schema data during encoding (#17565)
## Description

Currently, the schema data codec stringifies and string-parses the
schema data during encoding and decoding, respectively. While this is
required for the schema summarizer, it is not a requirement for encoding
schema data in general, so it has been removed from the codec and put
into the summarizer itself. This will allow other places in the system
to encode and decode schema data without automatic stringification.