build(docs): update throws and examples sections to new titles (#17616)
Consume the updated `api-markdown-documenter` package to allow custom
titles for the `@example` and `@throws` sections.

1. The `api-markdown-documenter` package has utility functions
`createExamplesSection` and `createThrowsSection` in `Helpers.ts` which
was updated to include a optional headingTitle param in v0.9.0. These
functions generate sections that describe the `@example` and `@throws`
comments on an API item, respectively.
   
2. At present, the headings "Examples" and "Throws" are hard-coded for
these sections. For FF.com, we aim to use custom titles "Usage" for the
`@example` section and "Error Handling" for the `@throws` section.
   
3. Adjustments on FF.com:
- The default title in the examples section should be changed from
"Examples" to "Usage".
- The default title in the throws section should be changed from
"Throws" to "Error Handling".


![image](https://github.com/microsoft/FluidFramework/assets/107130183/08b1bee0-5340-4c97-8239-36aab7f04372)

![image](https://github.com/microsoft/FluidFramework/assets/107130183/f61299bb-8c6b-4b66-950f-f7c46076bded)



[AB#5581](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5581)