short term functional approach to sampling telemetry with an example of usage in OpPerfTelemetry (#17075)
## Description
Adds an implementation of the short term functional approach to sampling
telemetry and applies it to a few places within the FF codebase. Note
that when we refer to "sampling" here, we mean: Emitting a subset of
total events. At this time, this is not meant to provide any form of
data aggregation from the unsampled events.

The primary goal of this is to:
1. Provie a short term solution for allowing customers to turn sampling
on or off (in this case, using the `Fluid.Telemetry.DisableSampling`
logger config value)
1. Streamline some of the commonly used sampling logic across the Fluid
codebase, which will also alleviate the need to manually track sampling
state logic.

This new simplistic logger utility, wraps around an existing logger
matching the `ITelemetryBaseLogger` or `ITelemetryLoggerExt` interface
and applies a provided callback to determine if an event should be
sampled in conjunction with the `Fluid.Telemetry.DisableSampling` logger
config value.

If `Fluid.Telemetry.DisableSampling` is set to true, all events will be
unsampled, otherwise they will be sampled according to your provided
event sampler callback.

### Detailed change list:
#### packages/runtime/container-runtime/src/connectionTelemetry.ts
1. Swaps delta latency telemetry events to using the new sampled logger.
1. Refactor of how Op latency metrics are emitted to be more intuitive
(the new sampled logger utility was NOT applied due to the complexity of
how op latency metrics are collected in an asynchronous manner.
#### experimental/dds/tree/src/SharedTree.ts
1. Swaps IdCompressor telemetry within SharedTree class to use the new
sampled logger utility.