build: Add per-feed package lists to aid reviews (#17634)
As we add/remove/move packages, these lists will change and we can use
them to verify _before we run the publishing pipeline_ what will be
published. The CI pipeline just publishes the packages in each list to
the feed in question. This makes it far less likely that we find out a
package is missing in a feed only after we do a release.

They are generated by running the `generate:packageList` script from the
root of the release group. They are also generated during a build, so
these files are kept up-to-date as changes are made.