add generic fuzz test minimization (#17636)
Run basic fuzz test minimization on all failing fuzz tests. 

This change more tightly couples fuzz test minimization to the actual
fuzz test running itself, as now that we have more options it becomes
necessary to update the minimization tool to also use those same options
(e.g. `containerRuntimeOptions` must be the same as the original test
run for minimization to be able to reproduce the failure).

The current implementation is a scaled down version of the current
implementation in `sequence`. It gets maybe 90% of the way there for all
test cases by running the deletion rule and ~50 random transforms. This
is sufficient to remove the majority of the noise, but the particulars
like the length of the created string may still be sub-optimal. Since
this runs on every failure, I wanted for the minimization here to be
less expensive.

For minifying tests to be used as unit or regression tests, I usually
run a couple thousand iterations of applying a transform as many times
as possible to each operation, somewhat skirting the pass ordering
problem -- since we have so few ops we can afford to run everything in a
for loop a bunch of times. To solve the case in which we want to
minimize a test case for a unit test, I want to add an option similar to
`skip` that lets one specify particular seeds to minimize fully.

That said, I think this serves as a sufficient baseline change that can
be built on, and also makes it easier for me to apply local changes that
I need to make some things work.