Don't ignore mock createBlob errors in blob manager tests (#17621)
Our mock createBlob was swallowing errors throw by itself. It hid a
couple of issues in the current tests so I improved abort tests,
ensuring only the statement that is expected to throw gets inside a try
block and making sure the error message is the one we expect. It also
hid some errors in getPendingBlobs tests caused by a recent change where
now the stashed blob process is done while connecting for the first
time.